+++ src/SerializerGenerator.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Liip\Serializer;
 
 use Liip\MetadataParser\Builder;
+use Liip\MetadataParser\Metadata\AbstractPropertyType;
 use Liip\MetadataParser\Metadata\ClassMetadata;
 use Liip\MetadataParser\Metadata\PropertyMetadata;
 use Liip\MetadataParser\Metadata\PropertyType;
@@ -202,23 +203,69 @@ final readonly class SerializerGenerator
 
         $modelPropertyPath = $modelPath.'->'.$propertyMetadata->getName();
         $fieldTarget = $target.'["'.$propertyMetadata->getSerializedName().'"]';
+        $type = $propertyMetadata->getType();
+        $requiresExplicitNullSet = $this->fieldTypeRequiresExplicitNullSet($type);
+        $shouldSerializeNull = $this->configuration->shouldSerializeNull();
+        $setNull = $shouldSerializeNull ? $this->templating->renderAssign($fieldTarget, 'null') : null;
 
         if ($propertyMetadata->getAccessor()->hasGetterMethod()) {
             $tempVariable = str_replace(['->', '[', ']', '$'], '', $modelPath).ucfirst($propertyMetadata->getName());
+            $value = $this->templating->renderGetter($modelPath, $propertyMetadata->getAccessor()->getGetterMethod());
+
+            if ($shouldSerializeNull && !$requiresExplicitNullSet) {
+                return $this->generateCodeForFieldType($type, $apiVersion, $serializerGroups, $fieldTarget, $value, $stack)."\n";
+            }
+
+            // the conditional is a null check, so the remaining expressions can assume non-null types
+            $nonNullType = ($type instanceof AbstractPropertyType) ? $type->asNullable(false) : $type;
 
             return $this->templating->renderConditional(
-                $this->templating->renderTempVariable($tempVariable, $this->templating->renderGetter($modelPath, $propertyMetadata->getAccessor()->getGetterMethod())),
-                $this->generateCodeForFieldType($propertyMetadata->getType(), $apiVersion, $serializerGroups, $fieldTarget, '$'.$tempVariable, $stack, $depth + 1)
+                $this->templating->renderTempVariable($tempVariable, $value),
+                $this->generateCodeForFieldType($nonNullType, $apiVersion, $serializerGroups, $fieldTarget, '$'.$tempVariable, $stack, $depth + 1),
+                $setNull
             );
         }
         if (!$propertyMetadata->isPublic()) {
             throw new \Exception(\sprintf('Property %s is not public and no getter has been defined. Stack %s', $modelPropertyPath, var_export($stack, true)));
         }
 
-        return $this->templating->renderConditional(
-            $modelPropertyPath,
-            $this->generateCodeForFieldType($propertyMetadata->getType(), $apiVersion, $serializerGroups, $fieldTarget, $modelPropertyPath, $stack, $depth + 1)
-        );
+        $serializeField = $this->generateCodeForFieldType($type, $apiVersion, $serializerGroups, $fieldTarget, $modelPropertyPath, $stack);
+
+        if (!$shouldSerializeNull) {
+            return $this->templating->renderConditional($modelPropertyPath, $serializeField);
+        }
+
+        return $requiresExplicitNullSet
+            ? $this->templating->renderConditional($modelPropertyPath, $serializeField, $setNull)
+            : "{$serializeField}\n";
+    }
+
+    /**
+     * Whether a PropertyType requires the target to be set to null in a separate statement (returns true), or its expression already allows the null case (returns false)
+     *
+     * @return bool True if the type needs a separate statement for the null case
+     */
+    private function fieldTypeRequiresExplicitNullSet(PropertyType $type): bool
+    {
+        if (!$type->isNullable()) {
+            return false;
+        }
+
+        switch ($type) {
+            case $type instanceof PropertyTypePrimitive:
+            case $type instanceof PropertyTypeUnknown:
+            case $type instanceof PropertyTypeDateTime: // can use the null-check operator
+            case $type instanceof PropertyTypeEnum: // can use the null-check operator
+                return false;
+
+            case $type instanceof PropertyTypeClass:
+            case $type instanceof PropertyTypeIterable:
+                return true;
+
+            case $type instanceof PropertyTypeUnion:
+        }
+
+        return false;
     }
 
     /**
@@ -237,11 +284,9 @@ final readonly class SerializerGenerator
         switch ($type) {
             case $type instanceof PropertyTypeDateTime:
                 $dateFormat = $type->getFormat() ?: \DateTimeInterface::ISO8601;
+                $dateToString = $this->templating->renderDateTime($modelPropertyPath, $dateFormat, $type->isNullable());
 
-                return $this->templating->renderAssign(
-                    $target,
-                    $this->templating->renderDateTime($modelPropertyPath, $dateFormat)
-                );
+                return $this->templating->renderAssign($target, $dateToString);
 
             case $type instanceof PropertyTypePrimitive:
             case $type instanceof PropertyTypeUnknown:
@@ -250,6 +295,7 @@ final readonly class SerializerGenerator
 
             case $type instanceof PropertyTypeEnum:
                 $valueAccess = $type->shouldSerializeAsValue() ? '->value' : '->name';
+                $valueAccess = ($type->isNullable() ? '?' : '').$valueAccess;
 
                 return $this->templating->renderAssign($target, $modelPropertyPath.$valueAccess);