+++ src/Template/Serialization.php
@@ -44,6 +44,11 @@ EOT;
 if (null !== {{condition}}) {
     {{code}}
 }
+{%- if elseCode -%}
+else {
+    {{elseCode}}
+}
+{% endif %}
 
 EOT;
 
@@ -102,7 +107,7 @@ EOT;
 
     private const string TMPL_GETTER = '{{modelPath}}->{{method}}()';
 
-    private const string TMPL_DATETIME = '{{propertyPath}}->format(\'{{format}}\')';
+    private const string TMPL_DATETIME = '{{propertyPath}}{{ nullable ? "?" : "" }}->format(\'{{format}}\')';
 
     private const string TMPL_TEMP_VAR = '${{name}} = {{value}}';
 
@@ -130,11 +135,12 @@ EOT;
         ]);
     }
 
-    public function renderConditional(string $condition, string $code): string
+    public function renderConditional(string $condition, string $code, ?string $elseCode = null): string
     {
         return $this->render(self::TMPL_CONDITIONAL, [
             'condition' => $condition,
             'code' => $code,
+            'elseCode' => $elseCode,
         ]);
     }
 
@@ -220,11 +226,12 @@ EOT;
         ]);
     }
 
-    public function renderDateTime(string $propertyPath, string $format): string
+    public function renderDateTime(string $propertyPath, string $format, bool $nullable = false): string
     {
         return $this->render(self::TMPL_DATETIME, [
             'propertyPath' => $propertyPath,
             'format' => $format,
+            'nullable' => $nullable,
         ]);
     }